refactor: 패키지를 MVC 형태로 분리
