+++ src/main/kotlin/blackjack/domain/GameStates.kt
@@ -1,4 +1,4 @@
-package blackjack
+package blackjack.domain
 
 enum class GameStates {
     END,
