+++ src/test/kotlin/blackjack/domain/CardNumberTests.kt
@@ -1,5 +1,6 @@
-package blackjack
+package blackjack.domain
 
+import blackjack.domain.CardNumber
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.CsvSource
