+++ src/test/kotlin/blackjack/domain/GameTest.kt
@@ -1,6 +1,6 @@
-package blackjack
+package blackjack.domain
 
-import blackjack.PlayFixture.TEST_NAME
+import blackjack.fixture.PlayFixture.TEST_NAME
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
